Implement working deletion with server.js
The only caveat is that two indexes, `last` and `contacts2`, depend on
flumeview-reduce, which seems to require a restart after
`flumeview.rebuild()` before indexing starts again.

See https://github.com/flumedb/flumeview-reduce/issues/10

Also, this throws a handful of "errors" when the rebuild request is
started, but they seem to be harmless. I'll add the errors as a comment
in the PR (#4).